Renamed TelemData folder and class to Pixhawk.
